Implement q_ascend and q_descend functions
First, reverse the list and find the minimum and maximum values.
While traversing the list, compare each traversed node with these values.
If the value of the traversed node is smaller or greater than the
minimum or maximum value respectively, remove the current node.
Once the traversal of the entire linked list is complete,reverse it again.